fix: bytes becoming message pack lists
In an attempt to make the terms become the smallest possible element in
message pack, I tried to render them as bytes. This resulted in a bytes
string instead of a single byte. This siginificantly increased the size
of the final encoded did so with the fix applied, we're even smaller.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>